pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-hana-retriever"
version = "0.1.0"
description = "LangChain BM25 and hybrid retrievers for SAP HANA Cloud"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "David Diaz"},
]
keywords = ["langchain", "retriever", "bm25", "hybrid", "sap", "hana", "hana-cloud", "btp", "rag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "langchain-core>=0.2.0",
    "hdbcli>=2.18.0",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
hybrid = [
    "langchain-community>=0.2.0",
]
dev = [
    "pytest>=7.0",
    "python-dotenv>=1.0",
]

[project.urls]
Homepage = "https://github.com/stubborncoder/langchain-hana-retriever"
Repository = "https://github.com/stubborncoder/langchain-hana-retriever"
Issues = "https://github.com/stubborncoder/langchain-hana-retriever/issues"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
